requirements.txt
nats-py==2.7.2
pyzmq==25.1.2
numpy==1.26.4
rpi-ws281x==5.0.0
